Merge old streams
Copy old buffer contents into the new buffer if we have to
re-initialize audio. This will allow proper gapless audio playback if
the speed is changed in real time. Well…. except for the "partial
remainder" rollover value used to alter the first output sample of the
next resampling block in the resampled, but that's extremely minor and
not worth complicating the fuck out of the codebase for.